+++ hevc/sps_test.go
@@ -377,3 +377,26 @@ func TestSPSParser3(t *testing.T) {
 		t.Errorf("Got %dx%d instead of %dx%d", gotWidth, gotHeight, expWidth, expHeight)
 	}
 }
+
+// TestParseSPSWithNonZeroNumDeltaPocs checks that parsing succeeds (Github issue #279)
+func TestParseSPSWithNonZeroNumDeltaPocs(t *testing.T) {
+	data := []byte{1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255,
+		3, 32, 0, 1, 0, 25, 64, 1, 12, 1, 255, 255, 1, 96, 0, 0, 3, 0, 0, 3, 0, 0, 3,
+		0, 0, 3, 0, 153, 53, 2, 64, 33, 0, 1, 0, 40, 66, 1, 1, 1, 96, 0, 0, 3, 0, 0, 3,
+		0, 0, 3, 0, 0, 3, 0, 153, 160, 2, 128, 128, 45, 22, 141, 82, 187, 34, 186, 173,
+		146, 169, 119, 53, 1, 1, 1, 0, 128, 34, 0, 1, 0, 8, 68, 1, 192, 36, 103, 192, 204, 100}
+	hevcd, err := DecodeHEVCDecConfRec(data)
+	if err != nil {
+		t.Error(err)
+	}
+
+	spsBytes := hevcd.GetNalusForType(NALU_SPS)
+	if len(spsBytes) != 1 {
+		t.Error("expected 1 sps NALU")
+	}
+	sps, err := ParseSPSNALUnit(spsBytes[0])
+	if err != nil {
+		t.Error(err)
+	}
+	t.Log(sps)
+}