+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- nothing yet
+
+## [0.39.0] - 2023-10-13
+
 ### Changed
 
 - TfraEntry Time and MoofOffset types changed to unsigned
@@ -465,7 +469,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New unique repo name: `mp4ff`
 
-[Unreleased]: https://github.com/Eyevinn/mp4ff/compare/v0.38.1...HEAD
+[Unreleased]: https://github.com/Eyevinn/mp4ff/compare/v0.39.0...HEAD
+[0.39.0]: https://github.com/Eyevinn/mp4ff/compare/v0.38.0...v0.39.0
 [0.38.1]: https://github.com/Eyevinn/mp4ff/compare/v0.37.0...v0.38.0
 [0.38.0]: https://github.com/Eyevinn/mp4ff/compare/v0.37.0...v0.38.0
 [0.37.0]: https://github.com/Eyevinn/mp4ff/compare/v0.36.0...v0.37.0